requirements.txt
python-telegram-bot==12.4
requests==2.23
retrying==1.3.3
